+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file. Be aware th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Fade Guild micro button alongside others
+
 ## [4.1.0] 2026-01-24
 ### Fixed
 - Settings should now save more reliably after a UI reload