+++ CHANGELOG.md
@@ -3,12 +3,12 @@ All notable changes to this project will be documented in this file. Be aware th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## Unreleased
 ### Added
 - The Buff Frame and Debuff Frame now support mouseover (closes [#13](https://github.com/AvyAddons/BlizzardBars_Mouseover/issues/13))
 
 ### Fixed
-- Fade Guild micro button alongside others
+- Fixed micro menu buttons remaining visible on login until hovered (closes [#17](https://github.com/AvyAddons/BlizzardBars_Mouseover/issues/17))
 
 ## [4.1.0] 2026-01-24
 ### Fixed