+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file. Be aware th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [4.2.0] 2026-01-25
 ### Added
 - The Buff Frame and Debuff Frame now support mouseover (closes [#13](https://github.com/AvyAddons/BlizzardBars_Mouseover/issues/13))